% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{AU_adj}
\alias{AU_adj}
\title{Adjacency matrix for focal and adjacent assemblages}
\format{
Matrix
}
\usage{
AU_adj
}
\description{
This is a square matrix containing information abount the presence-absence of each focal cell assemblage and its respective neighborhood assemblages, following the chessboard queen's case.
}
\keyword{datasets}
