% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{AU_grid}
\alias{AU_grid}
\title{Australia grid map}
\format{
Shapefile
}
\usage{
AU_grid
}
\description{
This is a gridded shapefile of Australia, with grid cells at a 0.5-degree resolution.
}
\keyword{datasets}
