% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpR_sensitivity.R
\name{CpR_sensitivity}
\alias{CpR_sensitivity}
\title{Title}
\usage{
CpR_sensitivity(
  tree,
  vec,
  mat = NULL,
  asb = NULL,
  rate = NULL,
  samp = 0,
  comp = "sorensen",
  method = "multisite",
  criteria = "my",
  ncor = 0
)
}
\arguments{
\item{ncor}{}
}
\description{
Title
}
