% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{pass_asb}
\alias{pass_asb}
\title{Presence-absence matrices for focal and adjacent assemblages}
\format{
List
}
\usage{
pass_asb
}
\description{
This is a list of presence-absence matrices for passeriformes from Australia. Each object within the list represents a passerine presence-absence matrix contaning the assemblage of a focal cell and its respective neighborhood assemblages.
}
\keyword{datasets}
