% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{pass_mat}
\alias{pass_mat}
\title{Presence-absence matrix of Australian passeriformes}
\format{
Matrix
}
\source{
\url{https://www.birdlife.org/}
}
\usage{
pass_mat
}
\description{
This is a presence-absence matrix for passeriformes obtained from Australian grid-cells. Each column represents a specific passeriforme species, while each row corresponds to an assemblage (or grid cell) within the Australian grid-map. All passerine range maps were sourced from the BirdLife platform.
}
\keyword{datasets}
