% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{pass_trees}
\alias{pass_trees}
\title{Phylogenetic tree of passerines from Australia}
\format{
List of phylogenetic trees of format phylo
}
\source{
\url{doi:10.1038/nature11631}
}
\usage{
pass_trees
}
\description{
This is a list contaning 100 phylogenetic trees of passerines from Australia. Each object within the list corresponds to a sampled phylogeny of class "phylo". The phylogenies are subsets obtained from the global bird phylogeny (https://birdtree.org/) made available by Jetz et al. (2012) in their published research, "The global diversity of birds in space and time".
}
\keyword{datasets}
