% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_pieces.R
\name{phylo_pieces}
\alias{phylo_pieces}
\title{Title}
\usage{
phylo_pieces(
  tree,
  n,
  criteria = "my",
  method = 1,
  timeSteps = FALSE,
  dropNodes = FALSE,
  returnTree = FALSE
)
}
\arguments{
\item{returnTree}{}
}
\description{
Title
}
