% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_phylo.R
\name{r_phylo}
\alias{r_phylo}
\title{Title}
\usage{
r_phylo(
  tree,
  n,
  mat,
  asb,
  index = NULL,
  comp = "sorensen",
  method = "multisite",
  criteria = "my",
  ncor = 0
)
}
\arguments{
\item{ncor}{}
}
\description{
Title
}
